fixes some fire bugs

+++ code/modules/1713/oil_fires.dm
@@ -144,6 +144,7 @@
 
 /obj/effect/burning_oil/proc/burningproc()
 	spawn(35)
+//this tile
 		for (var/mob/living/L in src.loc)
 			L.IgniteMob()
 			L.adjustFireLoss(rand(15,20))
@@ -152,28 +153,54 @@
 			spawn(50)
 				qdel(O)
 
+		for (var/obj/OB in src.loc)
+			if (prob(35) && !istype(OB, /obj/effect/decal/cleanable/blood/oil) && OB.flammable)
+				OB.fire_act(900)
+
+		for (var/obj/effect/burning_oil/OTH in src.loc)
+			if (OTH != src)
+				qdel(OTH)
+//burn floors
 		if (istype(get_turf(src), /turf/floor/plating/grass))
 			var/turf/T = get_turf(src)
 			T.ChangeTurf(/turf/floor/dirt/burned)
-
-		for (var/obj/OB in src.loc)
-			if (prob(35) && !istype(OB, /obj/effect/decal/cleanable/blood/oil) && OB.flammable)
-				OB.fire_act(700)
-
+		else if (istype(get_turf(src), /turf/floor/wood))
+			var/turf/T = get_turf(src)
+			T.ChangeTurf(/turf/floor/wood_broken)
+//bordering tiles
 		for (var/obj/effect/decal/cleanable/blood/OL in range(1, get_turf(src)))
-			if (istype(OL, /obj/effect/decal/cleanable/blood/oil))
-				if (prob(15))
-					new/obj/effect/burning_oil(OL.loc)
-			if (istype(OL, /obj/effect/decal/cleanable/blood/tracks) && OL.basecolor == "#A10808")
-				if (prob(15))
-					new/obj/effect/burning_oil(OL.loc)
+			var/dont = FALSE
+			for (var/obj/effect/burning_oil/BOI in OL.loc)
+				dont = TRUE
+			if (!dont)
+				if (istype(OL, /obj/effect/decal/cleanable/blood/oil))
+					if (prob(15))
+						new/obj/effect/burning_oil(OL.loc)
+				if (istype(OL, /obj/effect/decal/cleanable/blood/tracks) && OL.basecolor == "#A10808")
+					if (prob(15))
+						new/obj/effect/burning_oil(OL.loc)
 
 		for (var/turf/floor/plating/grass/GR in range(1, get_turf(src)))
-			if (prob(5))
-				var/obj/effect/burning_oil/NB = new/obj/effect/burning_oil(GR)
-				spawn(80)
-					qdel(src)
-					qdel(NB)
+			var/dont = FALSE
+			for (var/obj/effect/burning_oil/BOI in GR.loc)
+				dont = TRUE
+			if (!dont)
+				if (prob(5))
+					var/obj/effect/burning_oil/NB = new/obj/effect/burning_oil(GR)
+					spawn(80)
+						qdel(src)
+						qdel(NB)
+		for (var/turf/floor/wood/WF in range(1, get_turf(src)))
+			var/dont = FALSE
+			for (var/obj/effect/burning_oil/BOI in WF.loc)
+				dont = TRUE
+			if (!dont)
+				if (prob(15))
+					var/obj/effect/burning_oil/NB = new/obj/effect/burning_oil(WF)
+					spawn(80)
+						qdel(src)
+						qdel(NB)
+//remove duplicates
 		for (var/obj/effect/burning_oil/OTH in src.loc)
 			if (OTH != src)
 				qdel(OTH)
